Fixes #2785: Severe errors in rrdtool python lib can permanently block zope
git-svn-id: http://dev.zenoss.com/svnint/trunk/core@21967 f5df01ce-0320-0410-9ed9-ddcd13c732cc

+++ Products/ZenRRD/RenderServer.py
@@ -88,7 +88,7 @@ class GraphRenderThread(threading.Thread):
 
     def kill(self):
         self.isRunning = False
-        os.kill(self.proc.pid, signal.SIGQUIT)
+        if self.proc: os.kill(self.proc.pid, signal.SIGQUIT)
 
 
 class RenderServer(RRDToolItem):
